:recycle: Refacto for objects
