+++ src/functions/actions/carry.js
@@ -1,16 +1,11 @@
 import { messages, settings } from '../../variables'
 import { fill, inventory } from '../actions'
-import { getCurrentLocation } from '../locations'
-import { addObjectToInventory, isObjectInInventory } from '../inventory'
-import {
-  changeObjectState,
-  getObjectFromLocation,
-  getObjectsList,
-  getObjectFromCurrentLocation,
-  isObjectALiquid, getObject,
-} from '../objects'
-import { getAction } from './utils'
 import { cageTheBird, getTheBird } from '../bird'
+import { addObjectToInventory, isObjectInInventory } from '../inventory'
+import { getCurrentLocation } from '../locations'
+import { getObjectFromCurrentLocation, isObjectALiquid, updateObjectState } from '../object'
+import { getObjectsList } from '../objects'
+import { getAction } from './utils'
 
 export const carry = (object, actionName, instruction) => {
   const { inventoryLimit } = settings
@@ -25,7 +20,7 @@ export const carry = (object, actionName, instruction) => {
   if (!object && getObjectsList.length > 1) return messages.doWhat(instruction)
   if (!object && getObjectsList.length === 1) [obj] = getObjectsList()
 
-  if (object) obj = getObjectFromLocation(object)
+  if (object) obj = getObjectFromCurrentLocation(object)
   if (!obj) return messages.doWhat(instruction)
 
   // Already carrying
@@ -44,12 +39,12 @@ export const carry = (object, actionName, instruction) => {
   if (obj.name === 'bird') return getTheBird(obj)
 
   if (obj.name === 'bottle' && conditions.currentLocation.fluid) {
-    const bottleState = changeObjectState(obj.name, conditions.currentLocation.oily ? 'oilBottle' : 'waterBottle')
+    const bottleState = updateObjectState(obj.name, conditions.currentLocation.oily ? 'oilBottle' : 'waterBottle')
     addObjectToInventory(obj.name)
     return `${bottleState.change}\n${messages.okMan}`
   }
 
-  if (obj.name === 'cage' && getObjectFromLocation('bird')) return cageTheBird(obj, instruction)
+  if (obj.name === 'cage' && getObjectFromCurrentLocation('bird')) return cageTheBird(obj, instruction)
 
   addObjectToInventory(obj.name)
   return messages.okMan
