+++ src/functions/bird.js
@@ -1,7 +1,7 @@
-import { destroy } from './actions'
 import { messages } from '../variables'
+import { destroy } from './actions'
 import { addObjectToInventory, isObjectInInventory } from './inventory'
-import { changeObjectState, getObjectFromLocation } from './objects'
+import { updateObjectState, getObjectFromCurrentLocation } from './object'
 
 export const getTheBird = (bird) => {
   if (bird.currentState === 'birdForestUncaged') {
@@ -10,19 +10,19 @@ export const getTheBird = (bird) => {
   } else if (bird.currentState !== 'birdCaged') {
     if (!isObjectInInventory('cage')) return messages.cannotCarry
     if (isObjectInInventory('rod')) return messages.birdEvades
-    changeObjectState(bird, 'birdCaged')
+    updateObjectState(bird, 'birdCaged')
     addObjectToInventory('bird')
     return messages.okMan
   }
 }
 
 export const cageTheBird = (cage, instruction) => {
-  const bird = getObjectFromLocation('bird')
+  const bird = getObjectFromCurrentLocation('bird')
   if (!bird) return messages.doWhat(instruction)
 
   if (bird && bird.currentState !== 'birdCaged') {
     if (isObjectInInventory('rod')) return messages.birdEvades
-    changeObjectState(bird, 'birdCaged')
+    updateObjectState(bird, 'birdCaged')
     addObjectToInventory('bird')
     addObjectToInventory('cage')
     return messages.okMan
