+++ src/variables/actions.js
@@ -49,9 +49,10 @@ export const actions = [
     verbs: ['drink'],
   },
   {
-    name: 'drop',
+    // name: 'drop',
+    name: 'discard',
     message: messages.arentCarrying,
-    verbs: ['drop', 'relea', 'free', 'disca', 'dump'],
+    verbs: ['drop', 'relea', 'free', 'disca', 'dump', 'discard'],
   },
   {
     name: 'eat',
